+++ example/a.ts
@@ -1,12 +1,12 @@
 import { Elysia, t } from '../src'
 
-export const webSocketHandler = () => {
-	return new Elysia().ws('/ws', {
-		open(ws) {
-			console.log('WebSocket connected')
-		},
-		message(ws, message) {}
+const plugin = new Elysia()
+	.onAfterHandle({ as: 'scoped' }, (x) => {
+		console.log(x.response)
+		return 'From AfterHandle'
 	})
-}
 
-new Elysia().use(webSocketHandler).listen(3000)
+new Elysia()
+	.use(plugin)
+	.get('/', () => 'From Index')
+	.listen(4848)